add: generic class MarkovChain
It was an attempt to write generic MarkovChain class that
internally uses ForgorChain when no memory needed and
RemberChain when it should remember previous states.
Currently the solution is to pass sequence generic type Seq
and it's traits type SeqTraits. The latter provides splitting
sequence into tokens whether it's vector<T> or string.
Passing token type T and pair of iterators will be better.